requirements.txt
Django==4.2
djangorestframework-camel-case==1.4.2
djangorestframework==3.14.0
drf-yasg==1.21.5
model-bakery==1.11.0
psycopg2-binary==2.9.6
pytest==7.2.2
pytest-django==4.5.2